requirements.txt
amqp==5.3.1
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
billiard==4.2.0
Brotli==1.1.0
CacheControl==0.12.14
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deprecation==2.1.0
dj-database-url==2.2.0
Django==5.1.3
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-decouple==2.1
django-extensions==3.2.3
django-filter==24.3
django-model-utils==4.5.1
django-rest-framework==0.1.0
django-rest-passwordreset==1.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gitdb==4.0.11
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonpickle==3.0.4
jsonpointer==2.4
kombu==5.3.7
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.17.0
setuptools==73.0.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
squareup==38.2.0.20241017
stripe==9.3.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
